+++ ReservationManagementSystem.Application/Features/Rooms/Commands/DeleteRoom/DeleteRoomHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using MediatR;
+using ReservationManagementSystem.Application.Common.Errors;
 using ReservationManagementSystem.Application.Features.Rooms.Common;
 using ReservationManagementSystem.Application.Interfaces.Repositories;
 using ReservationManagementSystem.Application.Wrappers;
@@ -19,6 +20,12 @@ public sealed class DeleteRoomHandler : IRequestHandler<DeleteRoomRequest, Resul
 
     public async Task<Result<RoomResponse>> Handle(DeleteRoomRequest request, CancellationToken cancellationToken)
     {
+        var isInUse = await _roomRepository.IsRoomInUseAsync(request.Id);
+        if (isInUse)
+        {
+            return Result<RoomResponse>.Failure(ValidationError.ResourceInUse());
+        }
+
         var room = await _roomRepository.Delete(request.Id, cancellationToken);
 
         if (room is null)
